isolating createLock tests
